Add tcp_bridge to open grbl port 23 on Linux
Since LightBurn only talks to grbl on port 23, a grbl emulator script
would need admin rights to open that port. However, we don't want to
run python scripts with sudo, so we create a small compiled program
that only opens the port for us and bridges it to stdin/stdout.
Then we can give that program the CAP_NET_BIND_SERVICE capability so
that it bind port 23 without sudo.